+++ examples/example.js
@@ -17,11 +17,11 @@
 var nsapi = require("../lib/api.js");
 
 // TODO: Replace the user agent with your own
-var api = new nsapi.NsApi("<user agent>");
+var api = new nsapi.NsApi("Your nation's name");
 
 // The following is a simple example that retrieves the nation Auralia's full
 // name and prints it to the console.
-function example1() {
+function nationApiExample() {
     return api.nationRequest("Auralia", ["fullname"])
               .then(function(data) {
                   console.log(data["fullname"]);
@@ -30,19 +30,17 @@ function example1() {
 
 // The following is the same example as example 1, but notice how it completes
 // much faster because the previous request was cached!
-function example2() {
-    return api.nationRequest("Auralia", ["fullname"])
-              .then(function(data) {
-                  console.log(data["fullname"]);
-              });
+function cacheExample() {
+    return nationApiExample();
 }
 
 // The following example retrieves the delegate, founder, and list of nations
-// in the region of Catholic
-function example3() {
+// in the region of Catholic.
+function regionApiExample() {
     return api.regionRequest("Catholic", ["nations", "delegate", "founder"])
               .then(function(data) {
                   console.log("Region of Catholic");
+                  console.log();
                   console.log("Delegate: " + data["delegate"]);
                   console.log("Founder: " + data["founder"]);
                   console.log("Nations: " + data["nations"].split(":"));
@@ -50,24 +48,34 @@ function example3() {
 }
 
 // The following example retrieves the last 5 founding happening entries.
-function example4() {
+function worldApiExample() {
     return api.worldRequest(["happenings"],
                             {filter: "founding", limit: "5"})
               .then(function(data) {
                   for (var i = 0; i < data["happenings"]["event"].length; i++) {
                       var event = data["happenings"]["event"][i];
+                      console.log();
                       console.log("Event ID: " + event["id"]);
                       console.log("Event Timestamp: " + event["timestamp"]);
                       console.log("Event Text: " + event["text"]);
-                      console.log();
                   }
               })
+}
+
+// The following example retrieves information about the last Security Council
+// resolution at vote.
+function worldAssemblyApiExample() {
+    return api.worldAssemblyRequest(nsapi.WorldAssemblyCouncil.SecurityCouncil,
+                                    ["lastresolution"])
+              .then(function(data) {
+                  console.log(data["lastresolution"]);
+              });
 }
 
 // The following is a complex example that retrieves and sorts the list of
 // nations in the region of Catholic by their influence score, then prints the
 // list to the console.
-function example5() {
+function complexExample() {
     return api.regionRequest("Catholic", ["nations"])
               .then(function(data) {
                   var nations = data["nations"].split(":");
@@ -108,13 +116,40 @@ function example5() {
     }
 }
 
-// The following example uses private shards to retrieve the notices associated
-// with a nation from the last 24 hours and print them to the console, along
-// with the PIN required for future private shard requests.
-function example6() {
+function telegramExample() {
+    // TODO: Replace telegram details with your own
+    var clientKey = "";
+    var telegramId = "";
+    var telegramSecretKey = "";
+
+    return api.telegramRequest(clientKey, telegramId, telegramSecretKey,
+                               "Auralia", nsapi.TelegramType.NonRecruitment)
+              .then(() => console.log("Telegram sent"))
+              .catch(err => console.log("Telegram was not sent: " + err));
+}
+
+function authenticationExample() {
+    // TODO: Replace nation name and checksum with your own
+    var nation = "";
+    var checksum = "";
+
+    return api.authenticateRequest(nation, checksum)
+              .then(success => {
+                  if (success) {
+                      console.log("Authentication succeeded");
+                  } else {
+                      console.log("Authentication failed");
+                  }
+              });
+}
+
+// The following example uses private shards to retrieve the next issue time
+// and print it to the console, along with the PIN required for future private
+// shard requests.
+function privateShardsExample() {
     // TODO: Replace the nation name and password with your own
-    var nationName = "<nation name>";
-    var nationPassword = "<nation password>";
+    var nationName = "";
+    var nationPassword = "";
 
     let auth = {
         password: nationPassword,
@@ -122,11 +157,11 @@ function example6() {
     };
     return api.nationRequest(
         nationName,
-        ["notices"],
-        {"from": String(Math.floor(Date.now() / 1000) - (60 * 60 * 24))},
+        ["nextissuetime"],
+        undefined,
         auth)
               .then(function(data) {
-                  console.log(data["notices"]);
+                  console.log("Next issue time: "+ data["nextissuetime"]);
                   console.log("PIN: " + auth.pin);
               });
 }
@@ -134,28 +169,40 @@ function example6() {
 // The following code executes each example.
 Promise.resolve()
        .then(function() {
-           console.log("Example 1:\n");
-           return example1();
+           console.log("Nation API example:\n");
+           return nationApiExample();
        })
        .then(function() {
-           console.log("\nExample 2:\n");
-           return example2();
+           console.log("\nCache example:\n");
+           return cacheExample();
        })
        .then(function() {
-           console.log("\nExample 3:\n");
-           return example3();
+           console.log("\nRegion API example:\n");
+           return regionApiExample();
        })
        .then(function() {
-           console.log("\nExample 4:\n");
-           return example4();
+           console.log("\nWorld API example:");
+           return worldApiExample();
        })
        .then(function() {
-           console.log("\nExample 5:\n");
-           return example5();
+           console.log("\nWorld Assembly API example:\n");
+           return worldAssemblyApiExample();
        })
        .then(function() {
-           console.log("\nExample 6:\n");
-           return example6();
+           console.log("\nComplex example:\n");
+           return complexExample();
+       })
+       .then(function() {
+           console.log("\nTelegram example:\n");
+           return telegramExample();
+       })
+       .then(function() {
+           console.log("\nAuthentication example:\n");
+           return authenticationExample();
+       })
+       .then(function() {
+           console.log("\nPrivate shards example:\n");
+           return privateShardsExample();
        })
        .catch(function(err) {
            console.log(err);